+++ homework/hyunwoo/backend/db/todoItems.js
@@ -1,0 +1,42 @@
+export default [
+    {
+      id: 1,
+      userId: 1,
+      title: "베이직반 과제",
+      doneAt: null,
+      createdAt: new Date(),
+      updatedAt: null,
+    },
+    {
+      id: 2,
+      userId: 2,
+      title: "팀프로젝트",
+      doneAt: null,
+      createdAt: new Date(),
+      updatedAt: null,
+    },
+    {
+      id: 3,
+      userId: 3,
+      title: "알고리즘 코드카타",
+      doneAt: null,
+      createdAt: new Date(),
+      updatedAt: null,
+    },
+    {
+      id: 4,
+      userId: 4,
+      title: "Node.js 공부",
+      doneAt: null,
+      createdAt: new Date(),
+      updatedAt: null,
+    },
+    {
+      id: 5,
+      userId: 5,
+      title: "청소",
+      doneAt: null,
+      createdAt: new Date(),
+      updatedAt: null,
+    },
+  ];
